+++ src/core/sim_object/so_uid.h
@@ -15,7 +15,6 @@
 #ifndef CORE_SIM_OBJECT_SO_UID_H_
 #define CORE_SIM_OBJECT_SO_UID_H_
 
-#include <atomic>
 #include <limits>
 #include "core/util/root.h"
 
@@ -107,98 +106,6 @@ class SoUid {
   // TODO delete assignement operator and copy ctor
 
   BDM_CLASS_DEF_NV(SoUid, 1);
-};
-
-
-
-/// This class generates unique ids for simulation objects events satisfying the
-/// EventId invariant. Thread safe.
-class SoUidGenerator {
- public:
-  SoUidGenerator(const SoUidGenerator&) = delete;
-  SoUidGenerator() : counter_(0) {}
-
-  SoUid NewSoUid() { return SoUid(counter_++); }
-
-  // Returns the highest index that was used for a SoUid
-  uint64_t GetHighestIndex() const { return counter_; }
-
- private:
-  std::atomic<typename SoUid::Index_t> counter_;  //!
-  BDM_CLASS_DEF_NV(SoUidGenerator, 1);
-};
-
-template <typename TValue>
-class SoUidMap {
-  struct Iterator {
-
-    SoUidMap* map_;
-    uint64_t idx_;
-  };
-
-public:
-  SoUidMap() {}
-
-  SoUidMap(const SoUidMap& other)
-    : data_(other.data_), so_uid_reused_(other.so_uid_reused_) {}
-
-  SoUidMap(uint64_t initial_size) {
-    data_.resize(initial_size);
-    so_uid_reused_.resize(initial_size, SoUid::kReusedMax);
-  }
-
-  void resize(uint64_t new_size) {  // NOLINT
-    data_.resize(new_size);
-    so_uid_reused_.resize(new_size, SoUid::kReusedMax);
-  }
-
-  void clear() {  // NOLINT
-    for (auto& el: so_uid_reused_) {
-      el = SoUid::kReusedMax;
-    }
-  }
-
-  void ParallelClear() {
-    #pragma omp parallel for
-    for (uint64_t i = 0; i < data_.size(); ++i) {
-      so_uid_reused_[i] = SoUid::kReusedMax;
-    }
-  }
-
-  uint64_t size() const {  // NOLINT
-    return data_.size();
-  }
-
-  void Remove(const SoUid& key) {
-    if (key.GetIndex() >= data_.size()) {
-      return;
-    }
-    so_uid_reused_[key.GetIndex()] = SoUid::kReusedMax;
-  }
-
-  bool Contains(const SoUid& uid) const {
-    auto idx = uid.GetIndex();
-    if (idx >= data_.size()) {
-      return false;
-    }
-    return uid.GetReused() == so_uid_reused_[idx];
-  }
-
-  void Insert(const SoUid& uid, const TValue& value) {
-    auto idx = uid.GetIndex();
-    data_[idx] = value;
-    so_uid_reused_[idx] = uid.GetReused();
-  }
-
-  const TValue& operator[](const SoUid& key) const {
-    return data_[key.GetIndex()];
-  }
-
-  // find, erase, begin, end
-
-private:
-  std::vector<TValue> data_;
-  std::vector<typename SoUid::Reused_t> so_uid_reused_;
 };
 
 }  // namespace bdm
