+++ src/core/sim_object/so_uid_generator.cc
@@ -12,7 +12,7 @@
 //
 // -----------------------------------------------------------------------------
 
-#include "core/sim_object/so_uid.h"
+#include "core/sim_object/so_uid_generator.h"
 
 namespace bdm {
